+++ checks/apps/quantumespresso/quantumespresso_check.py
@@ -3,35 +3,50 @@
 #
 # SPDX-License-Identifier: BSD-3-Clause
 
+import pathlib
+import re
+import sys
+
 import reframe as rfm
 import reframe.utility.sanity as sn
 
+sys.path.append(str(pathlib.Path(__file__).parent.parent.parent / 'mixins'))
+from extra_launcher_options import ExtraLauncherOptionsMixin
+from sarus_extra_launcher_options import SarusExtraLauncherOptionsMixin
+from cuda_visible_devices_all import CudaVisibleDevicesAllMixin
 
-class QuantumESPRESSOCheck(rfm.RunOnlyRegressionTest):
-    scale = parameter(['small', 'large'])
-    variant = parameter(['maint', 'prod'])
-    modules = ['QuantumESPRESSO']
-    executable = 'pw.x'
-    executable_opts = ['-in', 'ausurf.in', '-pd', '.true.']
-    extra_resources = {
-        'switches': {
-            'num_switches': 1
+qe_tests = {
+    'Au-surf': {
+        'hohgant:cpu': {
+            'energy_reference': -11427.09017218,
+            'performance_reference': [{
+                'N': 1,    # number of nodes
+                'R': 32,   # ranks per node
+                'T': 4,    # threads per rank
+                'P': 55.4  # performance. here is time to solution
+                }
+            ]
+        },
+        'hohgant:nvgpu': {
+            'energy_reference': -11427.09017218,
+            'performance_reference': [{
+                'N': 1,    # number of nodes
+                'R': 4,    # ranks per node
+                'T': 16,   # threads per rank
+                'P': 55.4  # performance. here is time to solution
+                }
+            ]
         }
     }
-    strict_check = False
-    maintainers = ['LM']
-    tags = {'scs'}
+}
 
-    @run_after('init')
-    def set_prog_envs_and_tags(self):
-        if self.current_system.name in ['eiger', 'pilatus']:
-            self.valid_prog_environs = ['cpeIntel']
-        else:
-            self.valid_prog_environs = ['builtin']
-
-        self.tags |= {
-            'maintenance' if self.variant == 'maint' else 'production'
-        }
+class QuantumESPRESSOCheck(rfm.RunOnlyRegressionTest):
+    energy_tolerance = 1.0e-6
+    test_name = 'Au-surf'
+    executable = 'pw.x'
+    # TODO: tests should all have pw.in as input file
+    executable_opts = ['-in', 'ausurf.in']
+    maintainers = ['antonk']
 
     @sanity_function
     def assert_simulation_success(self):
@@ -48,106 +63,69 @@ class QuantumESPRESSOCheck(rfm.RunOnlyRegressionTest):
         return sn.extractsingle(r'electrons.+\s(?P<wtime>\S+)s WALL',
                                 self.stdout, 'wtime', float)
 
-
-@rfm.simple_test
-class QuantumESPRESSOCpuCheck(QuantumESPRESSOCheck):
-    valid_systems = ['daint:mc', 'eiger:mc', 'pilatus:mc']
-    energy_tolerance = 1.0e-6
-
     @run_after('init')
     def setup_test(self):
-        self.descr = (f'QuantumESPRESSO CPU check (version: {self.scale}, '
-                      f'{self.variant})')
-        if self.scale == 'small':
-            self.valid_systems += ['dom:mc']
-            self.energy_reference = -11427.09017218
-            if self.current_system.name in ['daint', 'dom']:
-                self.num_tasks = 216
-                self.num_tasks_per_node = 36
-            elif self.current_system.name in ['eiger', 'pilatus']:
-                self.num_tasks = 96
-                self.num_tasks_per_node = 16
-                self.num_cpus_per_task = 16
-                self.num_tasks_per_core = 1
-                self.env_vars = {
-                    'MPICH_OFI_STARTUP_CONNECT': 1,
-                    'OMP_NUM_THREADS': 8,
-                    'OMP_PLACES': 'cores',
-                    'OMP_PROC_BIND': 'close'
-                }
+        self.descr = (f'QuantumESPRESSO ground state SCF check')
+        self.env_vars = {
+            'OMP_NUM_THREADS': '$SLURM_CPUS_PER_TASK'
+        }
 
-        else:
-            self.energy_reference = -11427.09017152
-            if self.current_system.name in ['daint']:
-                self.num_tasks = 576
-                self.num_tasks_per_node = 36
-            elif self.current_system.name in ['eiger', 'pilatus']:
-                self.num_tasks = 256
-                self.num_tasks_per_node = 16
-                self.num_cpus_per_task = 16
-                self.num_tasks_per_core = 1
-                self.env_vars = {
-                    'MPICH_OFI_STARTUP_CONNECT': 1,
-                    'OMP_NUM_THREADS': 8,
-                    'OMP_PLACES': 'cores',
-                    'OMP_PROC_BIND': 'close'
-                }
+    @run_after('setup')
+    def setup_reference_dict(self):
+        self.ref_dict = (
+            qe_tests[self.test_name][self.current_partition.fullname]
+        )
+
+    @run_after('setup')
+    def setup_resources(self):
+        N = self.ref_dict['performance_reference'][0]['N']
+        R = self.ref_dict['performance_reference'][0]['R']
+        T = self.ref_dict['performance_reference'][0]['T']
+
+        self.num_tasks = N * R
+        self.num_cpus_per_task = T
+        self.num_tasks_per_node = R
+
+    @run_before('sanity')
+    def set_sanity_reference(self):
+        self.energy_reference = self.ref_dict['energy_reference']
 
     @run_before('performance')
-    def set_reference(self):
-        references = {
-            'small': {
-                'dom:mc': {'time': (109.814, None, 0.10, 's')},
-                'daint:mc': {'time': (98.475, None, 0.10, 's')},
-                'eiger:mc': {'time': (55.51, None, 0.10, 's')},
-                'pilatus:mc': {'time': (50.981, None, 0.10, 's')}
-            },
-            'large': {
-                'daint:mc': {'time': (125.648, None, 0.10, 's')},
-                'eiger:mc': {'time': (47.165, None, 0.10, 's')},
-                'pilatus:mc': {'time': (43.211, None, 0.10, 's')}
-            }
+    def set_performance_reference(self):
+        self.reference = {
+            '*': {'time': (self.ref_dict['performance_reference'][0]['P'],
+                           None, 0.10, 's')}
         }
-        self.reference = references[self.scale]
 
-    @run_before('run')
-    def set_task_distribution(self):
-        self.job.options = ['--distribution=block:block']
+@rfm.simple_test
+class SARUS_QuantumESPRESSOCheck(QuantumESPRESSOCheck, SarusExtraLauncherOptionsMixin, CudaVisibleDevicesAllMixin):
+    tags = {'production', 'sarus'}
+    container_image = variable(str, value='NULL')
+    valid_prog_environs = ['builtin']
+    valid_systems = ['+nvgpu +cpu']
 
-    @run_before('run')
-    def set_cpu_binding(self):
-        self.job.launcher.options = ['--cpu-bind=cores']
+    @run_after('setup')
+    def setup_container_platform(self):
+        # if container_image was provided then initialize a container execution
+        if self.container_image != 'NULL':
+            self.container_platform = 'Sarus'
+            self.container_platform.image = self.container_image
+            self.container_platform.with_mpi = False
+            self.container_platform.pull_image = False
+            self.container_platform.command = (
+                f'{self.executable} {" ".join(self.executable_opts)}'
+            )
+        #else stop here, container_image can't be empty
 
 
 @rfm.simple_test
-class QuantumESPRESSOGpuCheck(QuantumESPRESSOCheck):
-    valid_systems = ['daint:gpu']
-    num_gpus_per_node = 1
-    num_tasks_per_node = 1
-    num_cpus_per_task = 12
-    energy_tolerance = 1.0e-7
+class UENV_QuantumESPRESSOCheck(QuantumESPRESSOCheck, ExtraLauncherOptionsMixin, CudaVisibleDevicesAllMixin):
+    tags = {'production', 'uenv'}
+    valid_prog_environs = ['+quantum-espresso']
+    valid_systems = ['+nvgpu +cpu +uenv']
 
-    @run_after('init')
-    def setup_test(self):
-        self.descr = (f'QuantumESPRESSO GPU check (version: {self.scale}, '
-                      f'{self.variant})')
-        if self.scale == 'small':
-            self.valid_systems += ['dom:gpu']
-            self.num_tasks = 6
-            self.energy_reference = -11427.09017168
-        else:
-            self.num_tasks = 16
-            self.energy_reference = -11427.09017179
+    @run_after('setup')
+    def setup_uenv_modules(self):
+        # is self.modules needed here?
+        modules = ['quantum-espresso']
 
-    @run_before('performance')
-    def set_reference(self):
-        references = {
-            'small': {
-                'dom:gpu': {'time': (55.724, None, 0.10, 's')},
-                'daint:gpu': {'time': (57.349, None, 0.10, 's')}
-            },
-            'large': {
-                'daint:gpu': {'time': (37.138, None, 0.10, 's')}
-            }
-        }
-        self.reference = references[self.scale]
