--- uenv_checks/apps/quantumespresso/src/ausurf.in
@@ -1,145 +1,0 @@
-&CONTROL
-  title = ' DEISA pw benchmark ',
-  calculation = 'scf',
-  restart_mode = 'from_scratch',
-  tprnfor = .TRUE.,
-  prefix = 'ausurf',
-  pseudo_dir = './',
-  outdir = './out/'
-/
-
-&SYSTEM
-  ibrav = 8, 
-  celldm(1) = 38.7583, 
-  celldm(2) = 0.494393, 
-  celldm(3) = 1.569966, 
-  nat = 112,
-  ntyp = 1,
-  nbnd = 800, 
-  ecutwfc = 25.0
-  occupations='smearing', smearing='marzari-vanderbilt', degauss=0.05
-/
-
-&ELECTRONS
-    diagonalization='david'
-    mixing_beta = 0.7
-/
-
-ATOMIC_SPECIES
- AU  196.96  Au.pbe-nd-van.UPF  
-
-K_POINTS (gamma)
-
-ATOMIC_POSITIONS (angstrom)
-AU       29.285000       40.578999        7.173000
-AU       29.285000       35.511002        7.173000
-AU       32.214001       40.578999        7.173000
-AU       32.214001       35.511002        7.173000
-AU       35.141998       40.578999        7.173000
-AU       35.141998       35.511002        7.173000
-AU       38.070999       40.578999        7.173000
-AU       38.070999       35.511002        7.173000
-AU       40.999001       40.578999        7.173000
-AU       40.999001       35.511002        7.173000
-AU       43.928001       40.578999        7.173000
-AU       43.928001       35.511002        7.173000
-AU       46.855999       40.578999        7.173000
-AU       46.855999       35.511002        7.173000
-AU       29.285000       42.270000        4.782000
-AU       29.285000       37.202000        4.782000
-AU       32.214001       42.270000        4.782000
-AU       32.214001       37.202000        4.782000
-AU       35.141998       42.270000        4.782000
-AU       35.141998       37.202000        4.782000
-AU       38.070999       42.270000        4.782000
-AU       38.070999       37.202000        4.782000
-AU       40.999001       42.270000        4.782000
-AU       40.999001       37.202000        4.782000
-AU       43.928001       42.270000        4.782000
-AU       43.928001       37.202000        4.782000
-AU       46.855999       42.270000        4.782000
-AU       46.855999       37.202000        4.782000
-AU       30.749001       43.115002        7.173000
-AU       30.749001       38.047001        7.173000
-AU       33.678001       43.115002        7.173000
-AU       33.678001       38.047001        7.173000
-AU       36.605999       43.115002        7.173000
-AU       36.605999       38.047001        7.173000
-AU       39.535000       43.115002        7.173000
-AU       39.535000       38.047001        7.173000
-AU       42.464001       43.115002        7.173000
-AU       42.464001       38.047001        7.173000
-AU       45.391998       43.115002        7.173000
-AU       45.391998       38.047001        7.173000
-AU       48.320999       43.115002        7.173000
-AU       48.320999       38.047001        7.173000
-AU       30.749001       34.666000        4.782000
-AU       30.749001       39.737999        4.782000
-AU       33.678001       34.666000        4.782000
-AU       33.678001       39.737999        4.782000
-AU       36.605999       34.666000        4.782000
-AU       36.605999       39.737999        4.782000
-AU       39.535000       34.666000        4.782000
-AU       39.535000       39.737999        4.782000
-AU       42.464001       34.666000        4.782000
-AU       42.464001       39.737999        4.782000
-AU       45.391998       34.666000        4.782000
-AU       45.391998       39.737999        4.782000
-AU       48.320999       34.666000        4.782000
-AU       48.320999       39.737999        4.782000
-AU       29.285000       40.578999        0.000000
-AU       29.285000       35.511002        0.000000
-AU       32.214001       40.578999        0.000000
-AU       32.214001       35.511002        0.000000
-AU       35.141998       40.578999        0.000000
-AU       35.141998       35.511002        0.000000
-AU       38.070999       40.578999        0.000000
-AU       38.070999       35.511002        0.000000
-AU       40.999001       40.578999        0.000000
-AU       40.999001       35.511002        0.000000
-AU       43.928001       40.578999        0.000000
-AU       43.928001       35.511002        0.000000
-AU       46.855999       40.578999        0.000000
-AU       46.855999       35.511002        0.000000
-AU       30.749001       41.424000        2.391000
-AU       30.749001       36.355999        2.391000
-AU       33.678001       41.424000        2.391000
-AU       33.678001       36.355999        2.391000
-AU       36.605999       41.424000        2.391000
-AU       36.605999       36.355999        2.391000
-AU       39.535000       41.424000        2.391000
-AU       39.535000       36.355999        2.391000
-AU       42.464001       41.424000        2.391000
-AU       42.464001       36.355999        2.391000
-AU       45.391998       41.424000        2.391000
-AU       45.391998       36.355999        2.391000
-AU       48.320999       41.424000        2.391000
-AU       48.320999       36.355999        2.391000
-AU       29.285000       43.959999        2.391000
-AU       29.285000       38.893002        2.391000
-AU       32.214001       43.959999        2.391000
-AU       32.214001       38.893002        2.391000
-AU       35.141998       43.959999        2.391000
-AU       35.141998       38.893002        2.391000
-AU       38.070999       43.959999        2.391000
-AU       38.070999       38.893002        2.391000
-AU       40.999001       43.959999        2.391000
-AU       40.999001       38.893002        2.391000
-AU       43.928001       43.959999        2.391000
-AU       43.928001       38.893002        2.391000
-AU       46.855999       43.959999        2.391000
-AU       46.855999       38.893002        2.391000
-AU       30.749001       43.115002        0.000000
-AU       30.749001       38.047001        0.000000
-AU       33.678001       43.115002        0.000000
-AU       33.678001       38.047001        0.000000
-AU       36.605999       43.115002        0.000000
-AU       36.605999       38.047001        0.000000
-AU       39.535000       43.115002        0.000000
-AU       39.535000       38.047001        0.000000
-AU       42.464001       43.115002        0.000000
-AU       42.464001       38.047001        0.000000
-AU       45.391998       43.115002        0.000000
-AU       45.391998       38.047001        0.000000
-AU       48.320999       43.115002        0.000000
-AU       48.320999       38.047001        0.000000
